降级burstkit4j-0.10-beta4.1

+++ burstKit/src/main/java/burst/kit/entity/BurstTimestamp.java
@@ -24,7 +24,7 @@ public final class BurstTimestamp {
     /**
      * @return The burst timestamp (number of seconds since Burst epoch)
      */
-    public int getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
